release 12/02/2024

+++ src/main/java/vn/edu/likelion/front_ice/dto/response/challenger/NextLevelResponse.java
@@ -10,6 +10,6 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class NextLevelResponse {
-    int score;
+    Double score;
     String rank;
 }
